requirements.txt
pyserial
pandas
pyqtgraph
pyqt5